+++ hyx/ratelimit/managers.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from hyx.ratelimit.tokenbucket import TokenBucket
+from hyx.ratelimit.buckets import TokenBucket
 
 
 class RateLimiter:
@@ -14,10 +14,8 @@ class TokenBucketLimiter(RateLimiter):
     Replenish tokens as time passes on. If tokens are available, executions can be allowed.
     Otherwise, it's going to be rejected with RateLimitExceeded
     """
-    
-     __slots__ = (
-        "_token_bucket",
-    )
+
+    __slots__ = ("_token_bucket",)
 
     def __init__(self, max_executions: float, per_time_secs: float, bucket_size: Optional[float] = None) -> None:
         self._token_bucket = TokenBucket(max_executions, per_time_secs, bucket_size)
